% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_daily_app_data.R
\name{run_daily_app_data}
\alias{run_daily_app_data}
\title{Run all scripts in data-raw to update data for the app}
\usage{
run_daily_app_data()
}
\value{
no output
}
\description{
Run all scripts in data-raw to update data for the app
}
